requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-taggit==6.1.0
idna==3.10
pillow==11.3.0
python-decouple==3.8
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
